# Performance enhancements (fixes #9)