Using the benchmark of #9 (with a switch to `FIRTiled` for the second one), I now get this:
```
BenchmarkTools.Trial: 
  samples:          115
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  3.88 mb
  allocs estimate:  45
  minimum time:     39.40 ms (0.00% GC)
  median time:      40.57 ms (0.00% GC)
  mean time:        43.80 ms (0.18% GC)
  maximum time:     76.27 ms (0.00% GC)
BenchmarkTools.Trial: 
  samples:          404
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  3.91 mb
  allocs estimate:  533
  minimum time:     11.55 ms (0.00% GC)
  median time:      11.87 ms (0.00% GC)
  mean time:        12.38 ms (0.77% GC)
  maximum time:     23.65 ms (2.71% GC)
BenchmarkTools.Trial: 
  samples:          356
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  15.34 mb
  allocs estimate:  94
  minimum time:     12.66 ms (0.00% GC)
  median time:      13.42 ms (3.70% GC)
  mean time:        14.07 ms (2.44% GC)
  maximum time:     25.55 ms (2.20% GC)
BenchmarkTools.Trial: 
  samples:          99
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  11.56 mb
  allocs estimate:  39
  minimum time:     48.02 ms (0.00% GC)
  median time:      49.08 ms (0.00% GC)
  mean time:        50.55 ms (0.53% GC)
  maximum time:     78.06 ms (0.00% GC)
BenchmarkTools.Trial: 
  samples:          275
  evals/sample:     1
  time tolerance:   5.00%
  memory tolerance: 1.00%
  memory estimate:  23.02 mb
  allocs estimate:  78
  minimum time:     17.07 ms (2.95% GC)
  median time:      17.63 ms (2.96% GC)
  mean time:        18.18 ms (2.93% GC)
  maximum time:     35.66 ms (1.85% GC)
```
That means that the separable algorithm runs in 67% of the time (single-threaded), and the dense algorithm runs in 83% of the time.

Closes #9. This is a **significant** uglification of the internal code, unfortunately. It would be nice if julia/LLVM did all of this automatically. That might come with `@polly` support.