@tobig 2016-10-25
That sounds certainly interesting. Maybe @MatthiasJReisinger is interested in taking a look. Polly should be able to help with some tiling and maybe even registertiling.