@timholy 2016-10-25
Tiling would be amazing, but even something that was smart enough to optimize the iteration range and indexing offsets would let me delete a *lot* of annoying (and error-prone) code. What I mean is changing stuff like
```jl
for i in rng
    for j in rng2
        out[i] += A[i+offset1+j] * kernel[j+offset2]
    end
end
```
The other big piece is resolving https://github.com/JuliaLang/julia/issues/9080.