@tobig 2016-10-25
OK. What do you mean by 'optimizing'? What kind of transformation do you expect? If it is easy to provide a self-contained before/after piece of code we can probably check if transforming it is easy.