@timholy 2016-10-25
Like the transformations I do in this PR. Instead of putting an extra add in the inner loop, just shift the iteration bounds, and overall calculate the combination of iteration bounds and inner-loop offsets that results in the fewest arithmetic operations.