@timholy 2016-10-25
The canonical test code:
```jl
for I in CartesianRange(indices(out))
    @unsafe for J in CartesianRange(indices(kernel))
        out[I] += img[I+J] * kernel[J]
    end
end
```
where any combination of `out`, `img`, and `kernel` might be an [OffsetArray](https://github.com/alsam/OffsetArrays.jl).