@tlnagy 2016-10-27
Can confirm, this speeds everything up quite a bit. 🎉