@MatthiasJReisinger 2016-10-28
Sorry for my delay Tim! I must admit I have not yet been able to look at this change in more detail but I will try to figure out if Polly can help out here.