@timholy 2016-10-29
No hurry, in the sense that I'm already handling (at least most of) the issues on the julia side. I pinged you more to give you some sense of where I think your work might help make things better in the future.